@@ -1,39 +1,45 @@
 {
-    "name": "discordjs-bot",
+    "name": "discordts-bot",
     "version": "1.0.0",
     "description": "A simple Javascript Discord Bot",
     "main": "tsc & node main.js",
+    "type": "module",
     "dependencies": {
-        "discord.js": "^12.5.1",
-        "eslint": "^7.20.0",
-        "ffmpeg-static": "^4.2.7",
-        "node-fetch": "^2.6.1",
+        "@discordjs/builders": "^0.9.0",
+        "@discordjs/rest": "0.1.0-canary.0",
+        "@discordjs/voice": "^0.7.5",
+        "discord-api-types": "^0.25.2",
+        "discord.js": "^13.3.1",
+        "eslint": "^8.4.1",
+        "ffmpeg-static": "^4.4.0",
+        "libsodium-wrappers": "^0.7.9",
+        "node-fetch": "^3.1.0",
         "opusscript": "0.0.8",
         "winston": "^3.3.3",
         "youtube-search": "^1.1.6",
         "ytdl-core": "^4.9.1"
     },
     "devDependencies": {
-        "@tsconfig/node14": "^1.0.0",
-        "@types/node": "^14.14.31",
-        "@types/node-fetch": "^2.5.10",
-        "prettier": "^2.2.1",
-        "typescript": "^4.2.2"
+        "@tsconfig/node16": "^1.0.2",
+        "@types/node": "^16.11.13",
+        "prettier": "^2.5.1",
+        "typescript": "^4.5.4"
     },
     "scripts": {
-        "start": "tsc & node build/main.js",
-        "dev": "tsc & node build/testMain.js",
+        "start": "tsc & node --experimental-json-modules build/main.js",
+        "dev": "tsc & node --experimental-json-modules build/testMain.js",
         "prettier": "prettier --config .prettierrc src/**/*.ts --write",
-        "build": "tsc"
+        "build": "tsc",
+        "deploy": "node build/deployCommands.js"
     },
     "repository": {
         "type": "git",
-        "url": "git+https://github.com/Sahneeintopf/DiscordJs-Bot.git"
+        "url": "git+https://github.com/Sahneeintopf/DiscordTs-Bot.git"
     },
     "author": "Fabian Willner",
     "license": "ISC",
     "bugs": {
-        "url": "https://github.com/Sahneeintopf/DiscordJs-Bot/issues"
+        "url": "https://github.com/Sahneeintopf/DiscordTs-Bot/issues"
     },
-    "homepage": "https://github.com/Sahneeintopf/DiscordJs-Bot#readme"
+    "homepage": "https://github.com/Sahneeintopf/DiscordTs-Bot#readme"
 }
