@@ -1,14 +1,13 @@
-import { argumentWrapper } from "../../interfaces/wrapperObject";
 import * as Discord from "discord.js";
+import fetch from "node-fetch";
 
-module.exports = {
+export default {
     name: "cat",
     description: "random Cat Image",
-    async execute(message: Discord.Message, context: argumentWrapper) {
-        const fetch = require("node-fetch");
-        const { file } = await fetch(
-            "https://aws.random.cat/meow"
-        ).then((response) => response.json());
+    async execute(message: Discord.Message, args: string[]) {
+        const response = await fetch("https://aws.random.cat/meow");
+        // @ts-ignore
+        const { file } = await response.json();
         message.channel.send(file);
     },
 };
