@@ -1,16 +1,17 @@
-import { argumentWrapper } from "../../interfaces/wrapperObject";
 import * as Discord from "discord.js";
 
-module.exports = {
+export default {
     name: "kick",
     description: "Poggers",
-    execute(message: Discord.Message, context: argumentWrapper) {
+    execute(message: Discord.Message, args: string[]) {
         if (!message.mentions.users.size) {
             message.reply("you need to tag a user in order to kick them!");
             return;
         }
-        const taggedUser: Discord.User = message.mentions.users.first();
+        const taggedUser: Discord.User | undefined = message.mentions.users.first();
 
-        message.channel.send(`You wanted to kick: ${taggedUser.username}`);
+        if (taggedUser !== undefined){
+            message.channel.send(`You wanted to kick: ${taggedUser.username}`);
+        }
     },
 };
