@@ -1,14 +1,13 @@
-import { argumentWrapper } from "../../interfaces/wrapperObject";
 import * as Discord from "discord.js";
 
-module.exports = {
+export default {
     name: "servercommand",
     aliases: ["server"],
     description: "Displays some information about the server",
-    execute(message: Discord.Message, context: argumentWrapper) {
+    execute(message: Discord.Message, args: string[]) {
         message.channel.send(
-            `Server name: ${message.guild.name}\n` +
-                `Total members: ${message.guild.memberCount}`
+            `Server name: ${message.guild?.name}\n` +
+                `Total members: ${message.guild?.memberCount}`
         );
     },
 };
