@@ -1,22 +1,21 @@
-import { argumentWrapper } from "../../interfaces/wrapperObject";
 import * as Discord from "discord.js";
+import fetch from "node-fetch";
 
-module.exports = {
+export default {
     name: "urban",
     description: "urban dictionary search",
-    async execute(message: Discord.Message, context: argumentWrapper) {
-        const { args } = context;
-
-        if (!args.length) {
+    async execute(message: Discord.Message, args: string[]) {
+        if (!args?.length) {
             return message.channel.send("You need to supply a search term!");
         }
-        const fetch = require("node-fetch");
-        const querystring = require("querystring");
+        const querystring = await import("querystring");
         const query = querystring.stringify({ term: args.join(" ") });
-
-        const { list } = await fetch(
+        const response = await fetch(
             `https://api.urbandictionary.com/v0/define?${query}`
-        ).then((response) => response.json());
+        );
+        // @ts-ignore
+        const { list } = await response.json();
+
         if (!list.length) {
             return message.channel.send(
                 `No results found for **${args.join(" ")}**.`
