@@ -1,8 +1,8 @@
-import { Bot } from "./src/bot";
-import { token } from "./credentials.json";
+import { Bot } from "./src/bot.js";
+import credentials from "./credentials.json";
 
-function main(){
-    const bot: Bot = new Bot(token);
+function main() {
+    const bot: Bot = new Bot(credentials.token);
 }
 
 main();
