@@ -1,8 +1,7 @@
-import { argumentWrapper } from "./wrapperObject";
 import * as Discord from "discord.js";
 
 export interface command {
-    execute(message: Discord.Message, context: argumentWrapper): void;
+    execute(message: Discord.Message, args: string[]): void;
     name: string;
     description?: string;
     args?: boolean;
